backend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.6.0
certifi==2021.10.8
chardet==3.0.4
Django==4.0.1
djangorestframework==3.13.1
html5lib==1.0.1
idna==2.7
nameparser==0.5.6
numpy==1.22.0
pandas==1.3.5
PennSDK==1.8.9
python-dateutil==2.8.2
pytz==2021.3
requests==2.20.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.24.3
webencodings==0.5.1
